display_test/requirements.txt
# Display Test Requirements
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0  # For better emoji support

